+++ src/delay.rs
@@ -55,6 +55,19 @@ impl Delay {
         Delay { clocks, syst }
     }
 
+    /// Get access to the underlying register block.
+    ///
+    /// # Safety
+    ///
+    /// This function is not _memory_ unsafe per se, but does not guarantee
+    /// anything about assumptions of invariants made in this implementation.
+    ///
+    /// Changing specific options can lead to un-expected behavior and nothing
+    /// is guaranteed.
+    pub unsafe fn peripheral(&mut self) -> &mut SYST {
+        &mut self.syst
+    }
+
     /// Releases the system timer (SysTick) resource
     pub fn free(self) -> SYST {
         self.syst