+++ src/rtc.rs
@@ -23,7 +23,7 @@ pub enum Error {
 /// Real Time Clock peripheral
 pub struct Rtc {
     /// RTC Peripheral register definition
-    pub regs: RTC,
+    rtc: RTC,
 }
 
 #[cfg(feature = "defmt")]
@@ -47,7 +47,7 @@ impl Rtc {
     /// The `bypass` argument is `true` if you're using an external oscillator that
     /// doesn't connect to `OSC32_IN`, such as a MEMS resonator.
     pub fn new(
-        regs: RTC,
+        rtc: RTC,
         prediv_s: u16,
         prediv_a: u8,
         bypass: bool,
@@ -55,14 +55,14 @@ impl Rtc {
         bdcr: &mut BDCR,
         pwr: &mut PWR,
     ) -> Self {
-        let mut result = Self { regs };
+        let mut result = Self { rtc };
 
         enable_lse(bdcr, bypass);
         unlock(apb1, pwr);
         enable(bdcr);
         result.set_24h_fmt();
 
-        result.regs.prer.modify(|_, w| {
+        result.rtc.prer.modify(|_, w| {
             w.prediv_s().bits(prediv_s);
             w.prediv_a().bits(prediv_a)
         });
@@ -72,22 +72,35 @@ impl Rtc {
 
     /// Sets calendar clock to 24 hr format
     pub fn set_24h_fmt(&mut self) {
-        self.regs.cr.modify(|_, w| w.fmt().set_bit());
+        self.rtc.cr.modify(|_, w| w.fmt().set_bit());
     }
     /// Sets calendar clock to 12 hr format
     pub fn set_12h_fmt(&mut self) {
-        self.regs.cr.modify(|_, w| w.fmt().clear_bit());
+        self.rtc.cr.modify(|_, w| w.fmt().clear_bit());
     }
 
     /// Reads current hour format selection
     pub fn is_24h_fmt(&self) -> bool {
-        self.regs.cr.read().fmt().bit()
+        self.rtc.cr.read().fmt().bit()
+    }
+
+    /// Get access to the underlying register block.
+    ///
+    /// # Safety
+    ///
+    /// This function is not _memory_ unsafe per se, but does not guarantee
+    /// anything about assumptions of invariants made in this implementation.
+    ///
+    /// Changing specific options can lead to un-expected behavior and nothing
+    /// is guaranteed.
+    pub unsafe fn peripheral(&mut self) -> &mut RTC {
+        &mut self.rtc
     }
 
     /// Release the RTC peripheral
     pub fn free(self) -> RTC {
         // TODO(Sh3Rm4n): Disable peripheral before releasing it.
-        self.regs
+        self.rtc
     }
 
     /// As described in Section 27.3.7 in RM0316,
@@ -98,20 +111,20 @@ impl Rtc {
         F: FnMut(&mut RTC),
     {
         // Disable write protection
-        self.regs.wpr.write(|w| unsafe { w.bits(0xCA) });
-        self.regs.wpr.write(|w| unsafe { w.bits(0x53) });
+        self.rtc.wpr.write(|w| unsafe { w.bits(0xCA) });
+        self.rtc.wpr.write(|w| unsafe { w.bits(0x53) });
         // Enter init mode
-        let isr = self.regs.isr.read();
+        let isr = self.rtc.isr.read();
         if isr.initf().bit_is_clear() {
-            self.regs.isr.modify(|_, w| w.init().set_bit());
-            while self.regs.isr.read().initf().bit_is_clear() {}
+            self.rtc.isr.modify(|_, w| w.init().set_bit());
+            while self.rtc.isr.read().initf().bit_is_clear() {}
         }
         // Invoke closure
-        closure(&mut self.regs);
+        closure(&mut self.rtc);
         // Exit init mode
-        self.regs.isr.modify(|_, w| w.init().clear_bit());
+        self.rtc.isr.modify(|_, w| w.init().clear_bit());
         // wait for last write to be done
-        while !self.regs.isr.read().initf().bit_is_clear() {}
+        while !self.rtc.isr.read().initf().bit_is_clear() {}
     }
 }
 
@@ -125,7 +138,7 @@ impl Rtcc for Rtc {
         let (ht, hu) = bcd2_encode(time.hour())?;
         let (mnt, mnu) = bcd2_encode(time.minute())?;
         let (st, su) = bcd2_encode(time.second())?;
-        self.regs.tr.write(|w| {
+        self.rtc.tr.write(|w| {
             w.ht().bits(ht);
             w.hu().bits(hu);
             w.mnt().bits(mnt);
@@ -143,7 +156,7 @@ impl Rtcc for Rtc {
             return Err(Error::InvalidInputData);
         }
         let (st, su) = bcd2_encode(seconds as u32)?;
-        self.modify(|regs| regs.tr.modify(|_, w| w.st().bits(st).su().bits(su)));
+        self.modify(|rtc| rtc.tr.modify(|_, w| w.st().bits(st).su().bits(su)));
 
         Ok(())
     }
@@ -153,7 +166,7 @@ impl Rtcc for Rtc {
             return Err(Error::InvalidInputData);
         }
         let (mnt, mnu) = bcd2_encode(minutes as u32)?;
-        self.modify(|regs| regs.tr.modify(|_, w| w.mnt().bits(mnt).mnu().bits(mnu)));
+        self.modify(|rtc| rtc.tr.modify(|_, w| w.mnt().bits(mnt).mnu().bits(mnu)));
 
         Ok(())
     }
@@ -165,7 +178,7 @@ impl Rtcc for Rtc {
             Hours::AM(_h) | Hours::PM(_h) => self.set_12h_fmt(),
         }
 
-        self.regs.tr.modify(|_, w| w.ht().bits(ht).hu().bits(hu));
+        self.rtc.tr.modify(|_, w| w.ht().bits(ht).hu().bits(hu));
 
         Ok(())
     }
@@ -174,7 +187,7 @@ impl Rtcc for Rtc {
         if !(1..=7).contains(&weekday) {
             return Err(Error::InvalidInputData);
         }
-        self.modify(|regs| regs.dr.modify(|_, w| unsafe { w.wdu().bits(weekday) }));
+        self.modify(|rtc| rtc.dr.modify(|_, w| unsafe { w.wdu().bits(weekday) }));
 
         Ok(())
     }
@@ -184,7 +197,7 @@ impl Rtcc for Rtc {
             return Err(Error::InvalidInputData);
         }
         let (dt, du) = bcd2_encode(day as u32)?;
-        self.modify(|regs| regs.dr.modify(|_, w| w.dt().bits(dt).du().bits(du)));
+        self.modify(|rtc| rtc.dr.modify(|_, w| w.dt().bits(dt).du().bits(du)));
 
         Ok(())
     }
@@ -194,7 +207,7 @@ impl Rtcc for Rtc {
             return Err(Error::InvalidInputData);
         }
         let (mt, mu) = bcd2_encode(month as u32)?;
-        self.modify(|regs| regs.dr.modify(|_, w| w.mt().bit(mt > 0).mu().bits(mu)));
+        self.modify(|rtc| rtc.dr.modify(|_, w| w.mt().bit(mt > 0).mu().bits(mu)));
 
         Ok(())
     }
@@ -204,7 +217,7 @@ impl Rtcc for Rtc {
             return Err(Error::InvalidInputData);
         }
         let (yt, yu) = bcd2_encode(year as u32)?;
-        self.modify(|regs| regs.dr.modify(|_, w| w.yt().bits(yt).yu().bits(yu)));
+        self.modify(|rtc| rtc.dr.modify(|_, w| w.yt().bits(yt).yu().bits(yu)));
 
         Ok(())
     }
@@ -220,7 +233,7 @@ impl Rtcc for Rtc {
         let (mt, mu) = bcd2_encode(date.month())?;
         let (dt, du) = bcd2_encode(date.day())?;
 
-        self.regs.dr.write(|w| {
+        self.rtc.dr.write(|w| {
             w.dt().bits(dt);
             w.du().bits(du);
             w.mt().bit(mt > 0);
@@ -246,7 +259,7 @@ impl Rtcc for Rtc {
         let (mnt, mnu) = bcd2_encode(date.minute())?;
         let (st, su) = bcd2_encode(date.second())?;
 
-        self.regs.dr.write(|w| {
+        self.rtc.dr.write(|w| {
             w.dt().bits(dt);
             w.du().bits(du);
             w.mt().bit(mt > 0);
@@ -255,7 +268,7 @@ impl Rtcc for Rtc {
             w.yu().bits(yu)
         });
 
-        self.regs.tr.write(|w| {
+        self.rtc.tr.write(|w| {
             w.ht().bits(ht);
             w.hu().bits(hu);
             w.mnt().bits(mnt);
@@ -269,19 +282,19 @@ impl Rtcc for Rtc {
     }
 
     fn get_seconds(&mut self) -> Result<u8, Self::Error> {
-        let tr = self.regs.tr.read();
+        let tr = self.rtc.tr.read();
         let seconds = bcd2_decode(tr.st().bits(), tr.su().bits());
         Ok(seconds as u8)
     }
 
     fn get_minutes(&mut self) -> Result<u8, Self::Error> {
-        let tr = self.regs.tr.read();
+        let tr = self.rtc.tr.read();
         let minutes = bcd2_decode(tr.mnt().bits(), tr.mnu().bits());
         Ok(minutes as u8)
     }
 
     fn get_hours(&mut self) -> Result<Hours, Self::Error> {
-        let tr = self.regs.tr.read();
+        let tr = self.rtc.tr.read();
         let hours = bcd2_decode(tr.ht().bits(), tr.hu().bits());
         if self.is_24h_fmt() {
             return Ok(Hours::H24(hours as u8));
@@ -306,26 +319,26 @@ impl Rtcc for Rtc {
     }
 
     fn get_weekday(&mut self) -> Result<u8, Self::Error> {
-        let dr = self.regs.dr.read();
+        let dr = self.rtc.dr.read();
         let weekday = bcd2_decode(dr.wdu().bits(), 0x00);
         Ok(weekday as u8)
     }
 
     fn get_day(&mut self) -> Result<u8, Self::Error> {
-        let dr = self.regs.dr.read();
+        let dr = self.rtc.dr.read();
         let day = bcd2_decode(dr.dt().bits(), dr.du().bits());
         Ok(day as u8)
     }
 
     fn get_month(&mut self) -> Result<u8, Self::Error> {
-        let dr = self.regs.dr.read();
+        let dr = self.rtc.dr.read();
         let mt: u8 = if dr.mt().bit() { 1 } else { 0 };
         let month = bcd2_decode(mt, dr.mu().bits());
         Ok(month as u8)
     }
 
     fn get_year(&mut self) -> Result<u16, Self::Error> {
-        let dr = self.regs.dr.read();
+        let dr = self.rtc.dr.read();
         let year = bcd2_decode(dr.yt().bits(), dr.yu().bits());
         Ok(year as u16)
     }